requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
amqp==5.0.6
    # via kombu
argh==0.26.2
    # via -r requirements.in
asgiref==3.4.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==4.0.0
    # via aioredis
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via
    #   -r requirements.in
    #   django-celery-beat
cffi==1.15.0
    # via cryptography
channels==3.0.4
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.3.1
    # via -r requirements.in
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
constantly==15.1.0
    # via twisted
cryptography==35.0.0
    # via
    #   autobahn
    #   luxdb
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via
    #   -r requirements.in
    #   channels
django==3.2.9
    # via
    #   -r requirements.in
    #   channels
    #   django-celery-beat
    #   django-model-utils
    #   django-timezone-field
django-celery-beat==2.2.1
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-timezone-field==4.2.1
    # via django-celery-beat
gunicorn==20.1.0
    # via -r requirements.in
hiredis==2.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   hyperlink
    #   twisted
incremental==21.3.0
    # via twisted
kombu==5.2.0
    # via celery
luxdb==0.0.8
    # via -r requirements.in
msgpack==1.0.2
    # via channels-redis
numpy==1.21.3
    # via luxdb
pillow==8.4.0
    # via -r requirements.in
prompt-toolkit==3.0.22
    # via click-repl
psycopg2==2.9.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.20
    # via cffi
pyopenssl==21.0.0
    # via twisted
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python-memcached==1.59
    # via -r requirements.in
pytz==2021.3
    # via
    #   celery
    #   django
    #   django-timezone-field
pyyaml==6.0
    # via -r requirements.in
redis==3.5.3
    # via -r requirements.in
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   pyopenssl
    #   python-dateutil
    #   python-memcached
    #   service-identity
sqlparse==0.4.2
    # via django
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==3.10.0.2
    # via
    #   async-timeout
    #   twisted
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==2.1.6
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
